Tests: Reduce MySQL workaround
Apply that only on classes that need it, not globally.

This should heavily reduce amount of time needed to run test with
MySQL/MariaDB.
