+++ weblate/memory/tests.py
@@ -20,6 +20,7 @@
 import json
 from io import StringIO
 
+from django.conf import settings
 from django.core.management import call_command
 from django.core.management.base import CommandError
 from django.urls import reverse
@@ -47,6 +48,16 @@ def add_document():
 
 
 class MemoryModelTest(FixtureTestCase):
+    @classmethod
+    def _databases_support_transactions(cls):
+        # This is workaroud for MySQL as FULL TEXT index does not work
+        # well inside a transaction, so we avoid using transactions for
+        # tests. Otherwise we end up with no matches for the query.
+        # See https://dev.mysql.com/doc/refman/5.6/en/innodb-fulltext-index.html
+        if settings.DATABASES["default"]["ENGINE"] == "django.db.backends.mysql":
+            return False
+        return super()._databases_support_transactions()
+
     def test_machine(self):
         add_document()
         unit = self.get_unit()
