+++ weblate/trans/tests/test_autotranslate.py
@@ -19,6 +19,7 @@
 
 """Test for automatic translation."""
 
+from django.conf import settings
 from django.core.management import call_command
 from django.core.management.base import CommandError
 from django.urls import reverse
@@ -148,6 +149,16 @@ class AutoTranslationTest(ViewTestCase):
 
 
 class AutoTranslationMtTest(ViewTestCase):
+    @classmethod
+    def _databases_support_transactions(cls):
+        # This is workaroud for MySQL as FULL TEXT index does not work
+        # well inside a transaction, so we avoid using transactions for
+        # tests. Otherwise we end up with no matches for the query.
+        # See https://dev.mysql.com/doc/refman/5.6/en/innodb-fulltext-index.html
+        if settings.DATABASES["default"]["ENGINE"] == "django.db.backends.mysql":
+            return False
+        return super()._databases_support_transactions()
+
     def setUp(self):
         super().setUp()
         # Need extra power
